Add IdDistributor interface and add default interceptor implementation
Add an IdDistributor interface to make id distribution to each item flexible and extendable. This change also stores the id distributor instance to the item adapter. The default implementation is stored static inside the interface so the atomic long is shared with all item adapters to not collicate with an second item adapter instance.